backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
python-dotenv==1.0.1
google-generativeai==0.7.2
Pillow==10.3.0
opencv-python-headless==4.9.0.80
aiofiles==23.2.1
httpx==0.27.0
pydantic==2.7.4
pydantic-settings==2.3.4